refactor: reorganize rule and serialization
